+++ README.markdown
@@ -1,50 +1,64 @@
-## Summarize - A Ruby C wrapper for Open Text Summarizer
+Summarize - A Ruby C wrapper for Open Text Summarizer
+=====================================================
 
-## Installation
-
-### Easy
+Installation
+------------
 
     gem install summarize
 
-### Manual
 
-    git clone https://github.com/ssoper/summarize.git
-    cd summarize
-    rake build
-    gem build summarize.gemspec
-    gem install summarize-1.0.4.gem
+Usage
+-----
 
-## Usage
+    require 'summarize'
 
-The summarize method is added to File which you can use to summarize the contents of any plain text file
+    Summarize.summarize('text to summarize')
+
+
+By default it uses an English dictionary for summarizing but forty languages are supported. Pass in the valid ISO 639 language code to use one. A ratio (default is 25%) can also be passed in.
+
+    # Parse an article using Portuguese stemming rules with a ratio of 50%
+    Summarize.summarize('texto para sumariar', language: 'pt', ratio: 50)
+
+
+You can also use custom stemming rules
+    Summarize.summarize('text to summarize', dictionary: 'path/to/custom/dictionary')
+
+
+Topics can also be returned
+
+    # Returns an array with content first and topics second
+    content, topics = Summarize.summarize('text to summarize', topics: true)
+    # content => summarized text
+    # topics  => Comma-delimited string of topics
+
+
+### Usage (with monkey patch)
+
+    require 'summarize/ext_core'
+
+
+The summarize method is added to File and String:
 
     File.open('path/to/file').summarize
+
 
 Or use the String method
 
     "text to summarize".summarize
 
-By default it uses an English dictionary for summarizing but forty languages are supported. Pass in the valid ISO 639 language code to use one. A ratio (default is 25%) can also be passed in.
 
-    # Parse an article using Portuguese stemming rules with a ratio of 50%
-    "texto para sumariar".summarize(:language => 'pt', :ratio => 50)
+As it is a Delegate to Summarize.summarize, same parameters are accepted.
 
-You can also use custom stemming rules
 
-    "text to summarize".summarize(:dictionary => 'path/to/custom/dictionary')
-
-Topics can also be returned
-
-    # Returns an array with content first and topics second
-    content, topics = "text to summarize".summarize(:topics => true)
-    # content => summarized text
-    # topics  => Comma-delimited string of topics
-
-## Dependencies
+Dependencies
+------------
 
 You must have glib-2.0 and libxml-2.0 installed and properly configured.
 
-## Author
+
+Author
+------
 
 Gem written by Sean Soper ([@ssoper](http://twitter.com/ssoper))
 
@@ -52,7 +66,9 @@ The Open Text Summarizer library was written by Nadav Rotem and can be found at 
 
 Gem fixed to handle UTF-8 encoding by Yury Batenko ([@svenyurgensson](http://twitter.com/svenyurgensson))
 
-## License
+
+License
+-------
 
 Copyright (C) 2010 Sean Soper <sean.soper@gmail.com>
