+++ lib/summarize/core_ext.rb
@@ -1,0 +1,5 @@
+require 'summarize'
+
+require 'summarize/core_ext/string'
+
+require 'summarize/core_ext/file'
