+++ lib/summarize/languages.rb
@@ -1,0 +1,41 @@
+module Summarize
+	LANGUAGES = [
+    'bg', # Bulgarian
+    'ca', # Catalan
+    'cs', # Czech
+    'cy', # Welsh
+    'da', # Danish
+    'de', # German
+    'el', # Greek
+    'en', # English
+    'eo', # Esperanto
+    'es', # Spanish
+    'et', # Creole
+    'eu', # Basque
+    'fi', # Finnish
+    'fr', # French
+    'ga', # Irish
+    'gl', # Galician
+    'he', # Hebrew
+    'hu', # Hungarian
+    'ia', # Interlingua
+    'id', # Indonesian
+    'is', # Icelandic
+    'it', # Italian
+    'lv', # Latvian
+    'mi', # Maori
+    'ms', # Malay
+    'mt', # Maltese
+    'nl', # Dutch
+    'nn', # Norweigan
+    'pl', # Polish
+    'pt', # Portuguese
+    'ro', # Romanian
+    'ru', # Russian
+    'sv', # Swedish
+    'tl', # Tagalog
+    'tr', # Turkish
+    'uk', # Ukrainian
+    'yi'  # Yiddish
+  ]
+end
