+++ lib/summarize/version.rb
@@ -1,0 +1,3 @@
+module Summarize
+	VERSION = "1.0.4"
+end
